feat: Add ArithmethicException
0으로 나눌 경우 발생하는 문제를 해소하기 위해 예외처리 추가.
ArithmethicException의 경우 정수를 0으로 나눌 때 에러를 감지하기
때문에 클래스 Performer에서 인자를 전달해줄 때와 클래스
CalculateNum의 메소드 divNum()에서 받는 파라미터를 정수형으로
설정하였다.

+++ 4weeks/stringCalc/src/main/java/Performer.java
@@ -22,8 +22,8 @@ public class Performer {
                         Double.parseDouble(formula[operator + 1]));
             }
             if (formula[operator].equals("/")) {
-                result = CalculateNum.divNum(Double.parseDouble(formula[operator - 1]),
-                        Double.parseDouble(formula[operator + 1]));
+                result = CalculateNum.divNum(Integer.parseInt(formula[operator - 1]),
+                        Integer.parseInt(formula[operator + 1]));
             }
             formula[operator + 1] = Double.toString(result);// 연산자 1개, 피연산자 2개가 아닐경우 연산을 계속 진행
         }
